# Fix handling of Enums in Literal types