--- a/tests/test_structure_attrs.py
+++ b/tests/test_structure_attrs.py
@@ -1,4 +1,5 @@
 """Loading of attrs classes."""
+from enum import Enum
 from ipaddress import IPv4Address, IPv6Address, ip_address
 from typing import Union
 from unittest.mock import Mock
@@ -155,15 +156,44 @@ def test_structure_literal(converter_cls):
 
 @pytest.mark.skipif(is_py37, reason="Not supported on 3.7")
 @pytest.mark.parametrize("converter_cls", [Converter, GenConverter])
+def test_structure_literal_enum(converter_cls):
+    """Structuring a class with a literal field works."""
+    from typing import Literal
+
+    converter = converter_cls()
+
+    class Foo(Enum):
+        FOO = 1
+        BAR = 2
+
+    @define
+    class ClassWithLiteral:
+        literal_field: Literal[Foo.FOO] = Foo.FOO
+
+    assert converter.structure(
+        {"literal_field": 1}, ClassWithLiteral
+    ) == ClassWithLiteral(Foo.FOO)
+
+
+@pytest.mark.skipif(is_py37, reason="Not supported on 3.7")
+@pytest.mark.parametrize("converter_cls", [Converter, GenConverter])
 def test_structure_literal_multiple(converter_cls):
     """Structuring a class with a literal field works."""
     from typing import Literal
 
     converter = converter_cls()
 
+    class Foo(Enum):
+        FOO = 7
+        FOOFOO = 77
+
+    class Bar(int, Enum):
+        BAR = 8
+        BARBAR = 88
+
     @define
     class ClassWithLiteral:
-        literal_field: Literal[4, 5] = 4
+        literal_field: Literal[4, 5, Foo.FOO, Bar.BARBAR] = 4
 
     assert converter.structure(
         {"literal_field": 4}, ClassWithLiteral
@@ -171,6 +201,14 @@ def test_structure_literal_multiple(converter_cls):
     assert converter.structure(
         {"literal_field": 5}, ClassWithLiteral
     ) == ClassWithLiteral(5)
+
+    assert converter.structure(
+        {"literal_field": 7}, ClassWithLiteral
+    ) == ClassWithLiteral(Foo.FOO)
+
+    cwl = converter.structure({"literal_field": 88}, ClassWithLiteral)
+    assert cwl == ClassWithLiteral(Bar.BARBAR)
+    assert isinstance(cwl.literal_field, Bar)
 
 
 @pytest.mark.skipif(is_py37, reason="Not supported on 3.7")