@Tinche 2022-02-16
Cool, can you compare the performance of structuring `Literal[1]` using `pyperf timeit`?