@brakhane 2022-02-16
OK, I've run tests with `pyperf timeit -s 'import cattrs' -s 'from typing import Literal' 'cattrs.structure(1, Literal[1])'` and noticed I could optimize the code a bit by only creating the list of Enums if the naïve check fails, I've pushed a new version.

```
original: Mean +- std dev: 1.34 us +- 0.01 us
old vers: Mean +- std dev: 2.03 us +- 0.02 us
new vers: Mean +- std dev: 1.44 us +- 0.03 us
```
original is the "broken" version without proper Enum support. "old" and "new" refer to the old vs the new patch.

I've also noticed that cattrs currently does *not* handle nested Literals correctly, which is why the test on 3.8 failed. Therefore, I've removed the additional check for that that I added, and might make a separate fix for that later.