@brakhane 2022-02-16
I'm sorry, I had overoptimized and broken the resolution for Enums that also inherit from int or str.

The new version is a bit slower, but I doubt it matters in reality, and it also does the correct thing now, which is a plus ;)

```
Not supporting Enums: Mean +- std dev: 1.92 us +- 0.03 us
Supporting Enums:    Mean +- std dev: 1.46 us +- 0.06 us
```