@Tinche 2022-02-16
Here are my (adjusted) benchmarks:

```
$ pyperf timeit -s 'from cattrs import structure; from typing import Literal; l = Literal[1]' 'structure(1, l)'
main: 1.33 us +- 0.08 us
pr: 2.03 us +- 0.13 us
```
So, significant.

You might look into how to turn this from a hook into a hook factory, and produce special hooks for literals with primitives and literals with enums. I can provide support on this.