@brakhane 2022-02-16
> Here are my (adjusted) benchmarks:
> 
> ```
> $ pyperf timeit -s 'from cattrs import structure; from typing import Literal; l = Literal[1]' 'structure(1, l)'
> main: 1.33 us +- 0.08 us
> pr: 2.03 us +- 0.13 us
> ```
> 
> So, significant.

One year ago I would have agreed, but since watching https://www.youtube.com/watch?v=r-TLSBdHe1A I now understand that these kinds of measurements should be taken with a grain of salt. (I don't mean this in a "I know better" kind of way, I just want to highly recommend the talk since it was an eye opener for me)

> 
> You might look into how to turn this from a hook into a hook factory, and produce special hooks for literals with primitives and literals with enums. I can provide support on this.

I think cattrs should do the correct thing by default. It literally (pun intended) violates the Literal constraint:

```py
class Foo(int, enum.Enum):
    FOO = 1

@attrs.define
class Bar:
    foo: Literal[Foo.FOO]

bar = cattrs.structure({"a": 1}, Bar)
print(type(bar.foo)) # int
```