@Tinche 2022-02-16
I'm saying we can do the right thing quickly by using hook factories. ;)