@brakhane 2022-02-16
> I'm saying we can do the right thing quickly by using hook factories. ;)

Ah, I misunderstood then. :) I will take a look at it later.