@brakhane 2022-02-16
OK, that was easier than expected. New version is back at `Mean +- std dev: 1.45 us +- 0.03 u`