@Tinche 2022-02-18
Ah I had a different solution in mind but yours is actually better. Just add a changelog entry please.