@brakhane 2022-02-18
> Ah I had a different solution in mind but yours is actually better. Just add a changelog entry please.

BTW, may I ask what you had in mind? I'm still new to cattrs, and probably haven't grokked all details yet.