@Tinche 2022-03-17
Hello! I'm merging in stuff in preparation for a release, and it's this PR's turn ;) So please rebase and let's get it in.

To answer your question, I was thinking of creating a hook factory per literal seen, with a special code path for the simple case to make it fast. But your approach accomplishes the same thing but simpler, so good job.